refactor: Extract Game Loop into a Standalone Component
