+++ App.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import AppNavigation from './navigation/AppNavigation';
 import GameContextProvider from './store/GameContext';
-import SocialLoop from './components/game/SocialLoop';
+import GameLoop from './components/game/GameLoop';
 
 export default function App() {
     return (
         <GameContextProvider>
-            <SocialLoop />
+            <GameLoop />
             <NavigationContainer>
                 <AppNavigation />
             </NavigationContainer>
